requirements.txt
numpy==1.17.4
doki-Mowstyl>=1.3.0
